Rename terminal.rs to agent.rs for more intuitive module naming
- Move agent implementation from terminal.rs to agent.rs
- Update module declarations with clippy annotation to allow module inception
- Simplify imports using re-exported types from agent module
- Maintain clean API while using more descriptive filename
