+++ src/agent/mod.rs
@@ -1,9 +1,10 @@
+#[allow(clippy::module_inception)]
+mod agent;
 pub mod agents;
 pub mod execution;
 pub mod monitor;
-pub mod terminal;
 
 // Re-export for convenience
+pub use agent::{Agent, AgentStatus};
 pub use agents::Agents;
 pub use monitor::AgentMonitor;
-pub use terminal::{Agent, AgentStatus};
