+++ src/rule/timeout.rs
@@ -3,8 +3,8 @@ use std::sync::Arc;
 use tokio::time::Duration;
 
 use crate::agent;
+use crate::agent::Agent;
 use crate::agent::execution::execute_rule_action;
-use crate::agent::terminal::Agent;
 use crate::config::RuleConfig;
 
 use super::Monitor;
